requirements.txt
alembic==1.12.0
asttokens==2.2.1
backcall==0.2.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.3
colorama==0.4.6
cryptography==41.0.3
decorator==5.1.1
dnspython==2.4.2
email-validator==2.0.0.post2
et-xmlfile==1.1.0
executing==1.2.0
fillpdf==0.7.2
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
idna==3.4
ipdb==0.13.11
ipython==8.8.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
numpy==1.24.1
openpyxl==3.0.10
pandas==1.5.3
parso==0.8.3
pdf2image==1.16.3
pdfminer.six==20221105
pdfrw2==0.5.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
PyMuPDF==1.23.3
PyMuPDFb==1.23.3
PyPDF2==3.0.1
PySimpleGUI==4.60.5
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2022.7.1
pywin32==306
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.20
stack-data==0.6.2
tomli==2.0.1
traitlets==5.8.1
typing_extensions==4.7.1
urllib3==2.0.4
Wand==0.6.11
wcwidth==0.2.6
Werkzeug==2.3.7
WTForms==3.0.1
xlwings==0.30.0
